fix(circuits): clean up of data for proof generation
Without this fix the data that is tested in test cases is different than
the data that is being passed for generation and verification of the
proof in amm and also in deposit transaction.

With this commit the same data in the test case is passed for generation
and verification of the proof. Also fix to calling main function with
await.

Issue-URL: https://app.clickup.com/t/8693d4amz