refactor(crypto): enhance message field validation and assertion
The commit refactors the `crypto` module, specifically the message bit
length and maximum value checks. This change was necessary because the
previous assertion function was not adequately validating the values in
the context of Snark fields. This change also adds a new assertion of
the maximum bits for the field value.