fix(dapp): construct pub zAsset utxo from utxo message
Without this patch, users will not be able to send or withdraw any zZKP
utxos that are coming from prp conversion.

This is a problem because with the current logic that zZKP utxo from prp
conversion has no amount (the amount is public) and it was not handled
correctly.

This patch solves the problem by decrypting the utxo message in the
correct way and converting public zAsset into private zAsset.

Issue-URL: https://app.clickup.com/t/8693pb5kr