chore(contracts): folder structure
This patch change the files structure in the
contracts ws. It splits the v0 and v1 contracts
and add a `common` folder for Vault since it's a
common contract between v0 and v1.

The Other minor change is using the Poseidon Hasher
in the Merkle Tree contract placed at
`contracts/common/binaryTree/BinaryIncrementalUpdatableMerkleTree`.

This patch also removes the `IMockTokens` since we can simply
use IErc20, IErc721 and IErc1155 from OZ contracts.