chore(contracts): hardware-wallets dep removed
Without this patch, our project includes the
"hardware-wallets" dependency, which, despite being
experimental and relying on deprecated dependencies
(hw-transport-u2f, node-hid), was not actually utilized
in our codebase.

This is a problem because it restricts our ability to upgrade
our development tools and environments. Specifically, it
prevents us from adopting newer versions of Node.js
(18/20+), which are required for the latest versions of Vite
and potentially SWC.

This patch solves the problem by removing the
"hardware-wallets" dependency. This allows for greater
flexibility in our development environment and enables us to
upgrade to the latest tooling versions without compatibility
issues.

Issue-URL: https://app.clickup.com/t/85ztxbdx3