refactor(contracts): validation of pub keys before registration (2 of 2)
This patch implements validation of public keys before a key gets registered.
Missed validation represented a problem as users could register improper keys
thus spending unnecessary gas and spamming the storage of contracts.
The root spending public key and the reading public key of a user, as well as
the public key of a trust provider shall be a point of the Baby Jubjub curve
belonging to the subgroup of points.
The previous patch implemented necessary functions of the `BabyJubJub` library.
This patch updates the `ProvidersKeys` smart contract to make it pre-validate
trust providers public keys.