chore(contracts): ZAccountRegistry_Proxy workaround in deploy scripts
For yet to be investigated reasons the `function deploy` in the deploy scripts
fails to deploy the EIP173Proxy with the nickname 'ZAccountRegistry_Proxy', but
the proxy gets properly deployed under the nickname 'zAccountRegistry_Proxy'
(note the difference in the 1st character).
In order to make scripts running, as a temporary solution, this patch changes
the proxy's nickname to 'zAccountRegistry_Proxy'.