chore(contracts): 16b-add-provider-key deploy scripts switched off
This script should register the public key of a trust provider. However, due to
a reason yet to be investigated, the script fails, which results in the failure
of the local dev environment deployment.
Providing
- trust providers keys may be added to the local dev environment separately
- scripts probably fails due to a problem with processing of EIP-712 signatures
- EIP-712 support by the ProvidersKeys smart contract is to be refactored soon
as a temporary solution, in order to make scripts running, this patch switches
off the failing script.