chore(dapp): storybook update
Without this patch, our project's storybook is not
at the latest stable version that's compatible with
our current environment.

This is a problem because the inability to utilize the latest
Storybook features and improvements hampers our UI
development and testing capabilities, impacting the overall
quality and efficiency of our development process.

This patch solves the problem by updating Storybook to the
latest compatible version (v7), enhancing our development
environment within the constraints of Node 16 compatibility.
It's noted that an upgrade to Storybook v8 is not feasible at
this time due to our project's reliance on Node 16.