fix(contracts): update roots
This patch fixes the zZone and zNetworks registry
contracts to update the root after adding a leaf.

It also fixes a function name `addZone` in the
zZone registry.

The comments have been removed from the
contracts as they have been moved to the files
placed in `lib/data/staticTree` directory.