fix(circuits): name of template LessThanWhenEnabled is misleading
Without this merge request template LessThanWhenEnabled naming will be
slightly misleading.

This is a problem for the readability of the code.

This merge request solves the problem by changing the template name
which is much readable along with a comment.

Audit Report Info - 4.1.25 V-PANC-VUL-025:
Name of template LessThanWhenEnabled is misleading

Issue-URL: https://app.clickup.com/t/869470xcb