feat(contracts): lock/unlock token by fee master
This patch adds `TransactionChargesHandler.sol` that adds the methods to
the pantherPool as follows:
- The new method `adjustVaultAssetsAndUpdateTotalFeeMasterDebt()` to let
feeMaster ask for locking or unlocking tokens. It also lets pantherPool
keep track of the total debts to the feeMaster.

- Two new methods `accountFeesAndReturnProtocolFeeAndMiningReward()` and
`accountFeesAndReturnMiningReward()` to be executed as part of each tx
to account fees.

It fixes a bug in Vault so the `lockAsset` method now accepts the native
tokens.

It adds `ForestTrees` type that contains 3 tree addresses and is used
at constructor instead of passng 3 addresses separately. It aims to fix
the `stack too deep` error in the constructor function.

It removes the logic and validations regarding the charged and added ZKP
amount from the `ZAccountsRegistry::activateZAccount()`.