refactor(contracts): get balance from vault
This patch adds `BalanceViewer` abstract contract. It contains `view`
methods for getting the Vault balance for any kind of token type.

It also refactors TransferHelper library to send a staticCall when only
reading the target contract. It lets to mark these methods as `view`.

`IVault` does not anymore inherit `IEthEscrow`. Instead, the `Vault`
inherits both `IVault` and `IEthEscrow` separately. The reason is to
decrease the size of contracts that uses `Vault` methods since they only
need `IVault`, not `IEthEscrow`.