feat(contracts): account mining rewards
This patch adds `MiningRewards` abstract contract to account miner
rewards. The miners should executte `claimMiningRewardWithSignature` or
`claimMiningReward` to claim the rewards.

This patch also modifies the `FeeMaster::payOff()` to accept the amount
of debt that is requested to be claimed. This change is needed since
the FeeMaster only knows the BusTree total fees, so whenever a miner
claims the rewards, the BusTree needs to specify how much of its total
debt needs to be claimed from the FeeMaster.