feat(contracts): store net value of premium reward
This patch changes the `uint96 _rewardReserve` to
`int192 _netRewardReserve`.

The busTree contract must know how much rewards miners have demanded
from the protocol. If the value is less than 0, the demanded amount is
more than the supplied, so the protocol must be able to fulfill it.
If the `_netRewardReserve` is less than 0, then the protocol fees
collected by feeMaster can be used by busTree.
A new function to allocate protocol fees to busTree must be added
in both feeMaster and busTree in the next follow-up commits.