feat(contracts): modify and clanup fee master
This path fixes the issue when swapping tokens and native tokens.
Two new functions are introduced:
`convertNativeToWNative` and `convertWNativeToNative`.

The `convertNativeToWNative` converts the input amount of native to wEth
before swapping them for erc20 token and the `convertWNativeToNative`
converts weth to native after swapping erc20 token for weth.

The FeeMaster now accounts for the trust provider fees so that they can
be withdrawn by executing the `payOff()` by the trustProvider EOA
address.

The logic for distributions of protocol ZKP fees is drafted to split
the ZKPs between prpConverter, busTree and, panther treasury.

Some refactors have been done such as introducing the `TotalDebtHandler`
and `ProtocolFeeSwapper` to keep the FeeMaster cleaner.