chore(contracts): del the unused contracts
This patch deletes 3 contracts:
- PolygonPrpRewardMsgRelayer
- ProtocolRewardController
- ToPolygonZkpTokenAndPrpRewardMsgSender

These 3 contracts are responsible for realising tokens from the
vestingPool in the Ethereum Mainnet and bridging them to prpConvertor in
the Polygon network. These contracts are not needed anymore since this
approach will be changed. A new contract shall be introduced to lock all
the tokens at once and release them to the prpConvertor linearly.