test(contracts): update BlacklistedZAccountId test
This patch updates the unit tests of the BlacklistedZAccountIdRegistry
contract.

The old version of the contract used 254 bits of the Merkle tree leaf,
so the bits from 0 to 253 could be used, and 254th and 255th bits must
be empty.

In the new version, the contract only uses 253 bits (from 0 to 252) and,
the 253th, 254th and, 255th bits must remain empty.

The unit test is updated based on these new updates which has been
already added into the contract.