fix(contracts): match vault allowance and reserves
This patch increases the Vault allowance when adding ZKPs to the
PrpConversion facet.

The Vault must be able to transfer ZKPs from the PrpConversion when user
converts PRPs to ZKPs. Increasing allowance when adding ZKPs as reserves
is needed to make sure that Vault is always allowed to transfer ZKPs
from the PrpConversion.