deploy(contracts): update deployment scripts
This patch adds `TestnetZkpAndLinkDonatory` to the Mock contracts that
are supposed to be used as a faucet in the test networks.

It adds paymaster deposit and add stake configuration scripts.

It updates `Account` contract constructor parameters to feat recent
updates.

It updates the deployment script for Vault and the configuration
script for the zSwap facet.

The updated VK for busTree has also been added.