fix(contracts): fee master
This patch fixes, refactors, and adds documentation to the FeeMaster
smart contract as follows:

**Fixes:**

- `rebalanceDebt` method: This method is intended to convert protocol
fees collected in any token other than ZKP and Native into ZKP and/or
Native via Uniswap V3.
A validation has been added to ensure that the token being sold to
Uniswap is neither ZKP nor Native. However, if the token is wNative, the
function skips the conversion of wNative to Native (since Uniswap does
not handle Native <-> wNative swaps) and proceeds to the next
conversion, which is Native (i.e., wNative) to ZKP. A new event
`ProtocolFeeSwapped` is emitted after the swap is completed.

- `distributeProtocolZkpFees` method: This method has been extended to
retrieve protocol fees collected in ZKP from the Vault and distribute
them to the Treasury and PantherPool (i.e. `prpConversion` facet). After
distributing the tokens, it resets the protocol fees to zero.
Additionally, the FeeMaster no longer distributes any ZKP fees to the
PantherTrees.

- `accountFees` methods: A modifier has been added to both `accountFees`
methods so that only PantherPool can trigger them.

- `_tryInternalZkpToNativeConversion` method: This function did not
calculate the Paymaster ZKP fees correctly in cases where not all of the
`paymasterCompensationInZkp` was converted to Native.
The formula has been updated to:

`(unconvertedNativeAmount * paymasterCompensationInZkp) / outputNative;`

where outputNative is the amount of Native resulting from the ZKP to
Native conversion.
A new event has also been added:
```
event PaymasterCompensationConverted(
    uint256 paymasterDebtInZkp,
    uint256 paymasterDebtInNative
);
```

**Refactors:**

- `addPool` method: It has been removed. FeeMaster already has an
updatePool method, which will now be used for adding, enabling, and
disabling pools. A validation has been added to the internal _updatePool
method to ensure the pool address is not zero. The `PoolUpdated` event
has been extended to emit not only the pool address but also the
pool `key`.

- `_updateDebtForProtocol` method: This method exists in the
FeeAccountant contract, which is a child contract of FeeMaster. It now
uses the internal `_updateDebts` method to update the debt for the
protocol instead of directly updating the debts mapping.

- Interfaces added/refactored: The following interfaces have been added
or refactored with proper documentation about their logic: `IFeeMaster`,
`IPayOff`, `IProviderFeeDebt`.

The `IFeeMasterHelper` interface has been removed from the project, and
the `Paymaster.sol` has been updated accordingly since it was importing
the `IFeeMasterHelper` interface.