fix(contracts): bus tree and bus queue
This patch fixes a bug in the BusTree contract where it reads the
BusTree root from storage, which can be zero at the beginning. To avoid
retrieving a zero value, it now reads the root by calling the view
function `getBusTreeRoot`.

This patch alos refactors the BusQueue contract, the method for
receiving ZKP as mining reward reserves from FeeMaster has been deleted.
If any miner demands more rewards from the protocol but no reserve is
available, the miner's demand cannot be fulfilled.