fix(contracts): correct main transaction type
This commit resolves a bug in the `TransactionTypes` library, where a
specific (sub-)type of a Main transaction was incorrectly classified as
"internal" when the transaction represented both a deposit and a
withdrawal. As a result, the smart contract failed to properly handle
token transfers, processing the transaction as if the tokens had been
correctly deposited or withdrawn, despite not being received or sent.

The `seenKytMessageHashes` has been marked as public since this value
is read by KYC/KYT povider.