feat(subgraph): update event names and data source addresses
Renamed the "RootUpdated" event to "StaticRootUpdated" to better reflect
the specificity of the event in the StaticTree contract. This change was
applied across the ABI, schema, and handler functions. Additionally,
updated the contract addresses and start blocks for several data
sources, reflecting the new contract deployments on the Polygon network.