fix(contracts): handle weth token out in zswap
This patch refactors and updates the `ZSwap` facet and plugin contracts
to fix a bug found in the plugin smart contracts.
The `PluginData` structure is updated to include both token in and token
out types. To achieve this, the public input, which has the encoded
address and type, is directly used to create the `PluginData`.

The old `PluginData` was:

```
struct PluginData {
    address tokenIn;
    uint96 amountIn;
    address tokenOut;
    uint8 tokenType;
    bytes data;
}
```

The updated `PluginData` is:

```
struct PluginData {
    uint168 tokenInTypeAndAddress;
    uint168 tokenOutTypeAndAddress;
    uint96 amountIn;
    bytes data;
}
```

In the swap plugins, the token types are used to determine if the token
is native or not. Since WETH has an ERC20 token type, it will be
considered an ERC20 token and Native token can be detrmine by the type
that `PluginData` includes.

The constants pointing to the indexes of tokens in the
`MainPublicSignals.sol` and `ZSwapPublicSignals.sol` have also been
renamed for clarity. In `MainPublicSignals`, the constant
`MAIN_TOKEN_IND` is renamed to `MAIN_TOKEN_TYPE_AND_ADDRESS_IND`.
The same update is applied to `ZSwapPublicSignals`.