feat(contracts): add zaccount renewal
This patch adds a new `renewZAccount` method to the
`ZAccountsRegistration` facet.

The logic of this method is identical to the `activateZAccount` method,
except it uses its own `txType`, `TT_ZACCOUNT_RENEWAL`.

The `activateZAccount` method has been refactored to utilize common
logic shared with both `renewZAccount` and `activateZAccount`, without
altering its original functionality.

Additionally, a new helper method `isActivationOrReactivationOrRenewal`
has been added to the `TransactionTypes` library. This method is used
when fees are charged and the `txNote` event is emitted, as these
processes are the same for activation, reactivation, and renewal.