feat(dapp): renew account button
Without this patch, users have no direct way to renew their
zAccount within the wallet interface, making the renewal
process unclear and potentially inconvenient.

This is a problem because users might miss opportunities to
extend their zAccount validity or might need to rely on
external instructions, leading to a less seamless experience
within the dApp.

This patch solves the problem by introducing a "Renew"
button in the wallet dialog under the "Your wallet/Your
Panther account" section. The button allows users to initiate
the renewal process A pending state is displayed during the
renewal, and a success toast message confirms when the
renewal is complete.

Issue-URL: https://app.clickup.com/t/86960hg12