fix(contracts): rewrite txai root with zero
The function `ForestTree.addUtxosToBusQueueAndTaxiTree()` takes a
parameter called `numTaxiUtxos` which represents the number of UTXOs to
be inserted into the TAXI tree. In case this parameter is zero, the
current implementation of the function will rewrite the root of the
TaxiTree with a zero value.

This patch addes a validation to taxi tree to make sure utxo array
that will be added to taxi has non-zero length