fix(contracts): use 256 bits as pool key
When a new pool is created, the key for that pool is calculated by
taking the first 4 bytes from a hash of the two tokens addresses - the
token pair - that the pool will serve. This key is not
collision-resistant and could be exploited.

This patch fixes the issue by using 256 bits for pool key.