fix(contracts): use safe24 to cast zacc id
The `safe24()` helper method is used to cast the zAccount ID to `uint24`.
The `uint24` type can hold up to 16,777,216 unique values, which is
sufficient for Panther users. Since each user must go through the KYC
process, it is unlikely that users can bypass registration. Therefore,
the total number of users is not expected to exceed 16 million.