fix(contracts): silent overflow in zasset registry
The library function
`ZAssetEncodingUtils.encodeTokenIdRangeSizeWithScale()` does not
validate the scaleFactor value.
Since this function is used in the ZAssetRegistryV1 contract during the
ZAsset inclusion process, if the caller specifies a scale factor that is
greater than 19, the expression of scale will silently overflow, and the
corresponding ZAsset leaf will contain an incorrect scale value.

This patch fixes the issue by validating the `scaleFactor` to be less
than or equal to 19.