fix(contracts): validate voucher limit
The check in the require statement allows owner to pass a pair of
`_limit` and `_amount` values, such that the new limit will be less than
`rewardsGenerated`.

The `require` has been updated to:

`require(_limit >= rewardsGenerated + _amount);`

This check would ensure that the limit is large enough to cover at least
one more reward of the specified amount.