fix(contracts): accounting total released zkp
The `ZkpReserveController` contract releases some ZKP tokens linearly to
the pool based on the number of blocks elapsed.

In the current implementation, `scTotalReleased` is updated by the
`_scReleasable` amount, but it does not account for the case where the
contract does not have enough balance to release the specified amount of
tokens.

This patch fixes the contract to properly account for the total released
ZKP when `_releasable` exceeds the contract's balance.