fix(circuits): registration/renewal circuit does not validate KYC expiry
In the current implementation Since KYC certificates are not tracked on
the smart-contract level, there is a possibility of reusing the same KYC
certificate for all further account renewals, even if certificates
become expired.

With these changes we are adding constaint to the circuits to check the
KYC certificate provided is not expired.