fix(circuits): nInputs in ZeroPaddedInputChecker is under constrained
Because nInputs is not correctly validated to be less than max_nInputs,
the less than comparison can be circumvented allowing an array of inputs
which are not enforced to be zero padded as intended.

With these changes we are using Num2Bits to ensure that nInputs is
constrained to be within 252 bits.