refactor(contracts): getQuoteAmount
This patch renames the `getQuoteAmount` function to
`getTrustedPoolQuoteAmount`
It also refactors the function to accept `token0` and `token1` as
parameters. This change ensures that the correct token pair is used for
fee calculations and other logic dependent on the quote amount.
It's more effeciency since the function now uses the provided tokens
instead of querying the pool to get the tokens.

The `_updatePool` function has been updated to checks the token0 and
token1 addresses to ensure that the correct token pair is provided.
These tokens will be stored in the pool struct and used in the
`getTrustedPoolQuoteAmount` function.

the `getEnabledPoolAddress` has been renamed to
`getEnabledPoolOrRevert`. It also returns the Pool struct instead of
the pool address.

Since the token0 and token1 must be checked when registering a pool,
the WETH address must be used when adding a pool with Native token0 or
token1. So no NATIVE_ADDRESS is needed to be passed to the function when
registering pool by `updatePool` or reading the pool by
`getEnabledPoolOrRevert`.