fix(contracts): RequiredPrefund validation
This patch changes the RequiredPrefund validation to add the
`gasPriceMarkupPct` to the basefee, not `tx.gasprice + block.basefee`.
```
        uint256 toleratedGasPrice = _addExtraPct(
            block.basefee,
            config.gasPriceMarkupPct
        );
```

The idea is to limit the extra gas that can be asked on top of the
basefee, but since tx.gasPrice includes both base fee and tip and the
the tip can be arbitrary, it is not safe to use tx.gasPrice in the
toleratedGasPrice calculation.