feat(crypto): update key derivation logic
This patch updates how the panther wallet keypairs are derived from the
user signature. This archived by introducing a "key index" which is used
to control which keypair to derive at a given index. This give us the
ability to derive multiple keys without any dependency between them (as
was with the previous design) as well as make it easy to add or remove
keypairs without introducing breaking changes.

Issue-URL: https://app.clickup.com/t/8697w8q4w