chore(subgraph): Update contract addresses for internal canary env
The environment variables in `.env.canary.internal` for the
subgraph have been updated. This includes addresses and start
blocks for the following contracts on the matic network:
- PANTHER_POOL
- PANTHER_FOREST
- ZKP_RESERVE_CONTROLLER
- VAULT_V1
- FEE_MASTER

These updates are necessary to point the canary subgraph deployment
to the correct contract instances, likely for a new deployment or
environment update on the matic network.